+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## CHANGELOG
 
+### v2.1.0
+MCM Integration added
+Added the ability to automatically load a round to the chamber with Shift + R. The same Shift+R removes the bullet again (If you do this with a loaded magazine the gun won't shoot keep in mind)
+fixed bug that it was not longer possible to clear chamber manually when a mag is attached to the weapon.
+
 ### 2.0.3
 improved overall code structure
 bugfix calling close() while being in the inventory while putting a round into the chamber no longer causes a softlock